Rename DI extentions methods and move it to related DI namespace
